+++ zarrs/src/array/array_dlpack_ext.rs
@@ -54,6 +54,14 @@ macro_rules! unsupported_dtypes {
             | DataType::RawBits(_)
             | DataType::String
             | DataType::Bytes
+            | DataType::NumpyDateTime64 {
+                unit: _,
+                scale_factor: _,
+            }
+            | DataType::NumpyTimeDelta64 {
+                unit: _,
+                scale_factor: _,
+            }
             | DataType::Extension(_)
     };
 }