+++ zarrs_registry/src/data_type.rs
@@ -105,3 +105,9 @@ pub const STRING: &str = "string";
 
 /// Unique identifier for the `bytes` data type (registered).
 pub const BYTES: &str = "bytes";
+
+/// Unique identifier for the `numpy.datetime64` data type (registered)
+pub const NUMPY_DATETIME64: &str = "numpy.datetime64";
+
+/// Unique identifier for the `numpy.timedelta64` data type (registered)
+pub const NUMPY_TIMEDELTA64: &str = "numpy.timedelta64";