+++ zarrs_metadata_ext/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `data_type::Numpy{DateTime64,TimeDelta64}DataTypeConfiguration`
+
 ## [0.1.0] - 2025-05-16
 
 ### Added