+++ zarrs_registry/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `data_type::{NUMPY_DATETIME64,NUMPY_TIMEDELTA64}`
+
 ## [0.1.2] - 2025-06-08
 
 ### Added